shnoor-meetings-backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pydantic
psycopg2-binary
